`srb init` prompts before overwriting existing config (#1005)

+++ gems/sorbet/lib/create-config.rb
@@ -2,7 +2,7 @@
 # frozen_string_literal: true
 
 require_relative './step_interface'
-
+require_relative './user_input'
 require 'fileutils'
 
 class Sorbet; end
@@ -11,14 +11,17 @@ class Sorbet::Private::CreateConfig
   SORBET_DIR = 'sorbet'
   SORBET_CONFIG_FILE = "#{SORBET_DIR}/config"
 
+  extend Sorbet::Private::UserInput
   include Sorbet::Private::StepInterface
 
   def self.main
     FileUtils.mkdir_p(SORBET_DIR)
 
-    File.open(SORBET_CONFIG_FILE, 'w') do |f|
-      f.puts('--dir')
-      f.puts('.')
+    if File.exist?(SORBET_CONFIG_FILE) && prompt_for_confirmation!("#{SORBET_CONFIG_FILE} already exists. Overwrite?")
+      File.open(SORBET_CONFIG_FILE, 'w') do |f|
+        f.puts('--dir')
+        f.puts('.')
+      end
     end
   end
